Renamed Validation message pattern
